Rename taskmanager server and taskmanager interface
